requirements.txt
asgiref==3.8.1
black==24.4.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.1.0
Django==5.0.4
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==22.0.0
h11==0.14.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
uvicorn==0.29.0
whitenoise==6.6.0
